Update recivide.conf
When recivide looks for a ban in the fail2ban log file, the ip becomes baned, but is already banned by the precedent filter. Then if you want to block a repeater offender, you should block it when you unban that ip.
So i propose to change the regex to Unban instead of Ban.